refactor(frontend): refactor results table
Consolidate the various table components into a single `ResultsTable` component, which can be used with any combination of columns and rows.
